ERR_PROXY_CONNECTION_FAILED bug fixed
Whenever the browser lost its network connection and a page was
refreshed during that outage, a network connection error would fire from
the tab and be logged within the onerroroccured file. This file would
then try to look up the tab that fired the refresh and try to point the
tab to our error page. Adding an extra check to see if the tab id match
the tabId in the details object sent to the error handler fixes a bug
that originally updated all tabs to point to the same url rather than
their own originating url.