Added link to Chrome settings bug
Added a link to the Chrome settings bug to the comments in the
settingsmanager.js file.

This bug apparently allows access to both microphone and location if the
camera setting is ever enabled, even when the other two are explicitly
disabled. The fix for this bug on our side is to reapply all settings
whenever any particular setting is "cleared."

This "reapply" method will be kept around for at least 5 versions to
ensure that user using the older version of chrome are still protected.